refactor: add `garble` and `evaluate` mod
[ci skip]
